+++ packages/twenty-front/src/modules/settings/page-layout/hooks/useEndPageLayoutDragSelection.ts
@@ -3,7 +3,6 @@ import { CommandMenuPages } from '@/command-menu/types/CommandMenuPages';
 import { useRecoilCallback } from 'recoil';
 import { isDefined } from 'twenty-shared/utils';
 import { IconAppWindow } from 'twenty-ui/display';
-import { pageLayoutCurrentTabIdForCreationState } from '../states/pageLayoutCurrentTabIdForCreation';
 import { pageLayoutDraggedAreaState } from '../states/pageLayoutDraggedAreaState';
 import { pageLayoutSelectedCellsState } from '../states/pageLayoutSelectedCellsState';
 import { calculateGridBoundsFromSelectedCells } from '../utils/calculateGridBoundsFromSelectedCells';
@@ -13,7 +12,7 @@ export const useEndPageLayoutDragSelection = () => {
 
   const endPageLayoutDragSelection = useRecoilCallback(
     ({ snapshot, set }) =>
-      (activeTabId?: string | null) => {
+      () => {
         const pageLayoutSelectedCells = snapshot
           .getLoadable(pageLayoutSelectedCellsState)
           .getValue();
@@ -25,10 +24,6 @@ export const useEndPageLayoutDragSelection = () => {
 
           if (isDefined(draggedBounds)) {
             set(pageLayoutDraggedAreaState, draggedBounds);
-
-            if (isDefined(activeTabId)) {
-              set(pageLayoutCurrentTabIdForCreationState, activeTabId);
-            }
 
             navigateCommandMenu({
               page: CommandMenuPages.PageLayoutWidgetTypeSelect,
