--- packages/twenty-front/src/modules/settings/page-layout/states/pageLayoutCurrentTabIdForCreation.ts
@@ -1,8 +1,0 @@
-import { createState } from 'twenty-ui/utilities';
-
-export const pageLayoutCurrentTabIdForCreationState = createState<
-  string | null
->({
-  key: 'pageLayoutCurrentTabIdForCreationState',
-  defaultValue: null,
-});
